Improve application mutator with types
